refactor(test): move _resolve_current_participant_state tests to dedicated module
Address PR #490 review feedback:
- Move SvcAddParticipantStatusUseCase unit tests out of the already-large
  test_service.py into test_svc_add_participant_status.py
- Fix FakeParticipantNoStatuses to use an instance-level list (was a
  mutable class-level list that could leak state between tests)
- Remove now-unused CS_vfd import from test_service.py
